+++ src/Parallel.cpp
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright 2011 Matthew Tytel
  *
  * dfex is free software: you can redistribute it and/or modify
@@ -24,7 +24,7 @@ Class Parallel::cls(string("Parallel"), newInstance);
 void Parallel::process(const sample_t* in, sample_t* out, int num) {
     memset(out, 0, num * sizeof(sample_t));
 
-    for (uint i = 0; i < mProcessors.size(); i++) {
+    for (size_t i = 0; i < mProcessors.size(); i++) {
         mProcessors[i]->process(in, mBuffer, num);
         Process::combine(mBuffer, out, out, num);
     }
